Refactor NeoForgedRepositoryFilter: update package structure and simplify filter method
